% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_hyper_init.R
\name{map_hyperprior_elicitation}
\alias{map_hyperprior_elicitation}
\title{Estimate the approximation made in the hyperprior elicitation for the number 
of predictors associated with each response by Monte-Carlo simulation.}
\usage{
map_hyperprior_elicitation(p0, p, q, n_draws = 1e+05, n_cpus = 1)
}
\arguments{
\item{p0}{Vector of size 2 whose entries are the prior expectation and 
variance of the number of predictors associated with each response.}

\item{p}{Number of candidate predictors.}

\item{q}{Number of responses.}

\item{n_draws}{Number of draws used for Monte-Carlo simulation (default 1e5).}

\item{n_cpus}{Number of cores used (default serial).}
}
\value{
A list of errors for the prior mean and prior standard deviation of 
  the number of predictors associated with each response.
}
\description{
This function is used to provide guide hyperprior elicitation before running
\code{\link{atlasqtl}}.
}
\examples{
seed <- 123; set.seed(seed)
n <- 200; p <- 100; q <- 20000
p0 <- c(1, 10)
n_draws <- 1e5 # must be large for accurate estimation

map_hyperprior_elicitation(p0, p, q, n_draws, n_cpus = 1)

}
\seealso{
\code{\link{atlasqtl}}
}
