% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_output.R
\name{plot.atlasqtl}
\alias{plot.atlasqtl}
\title{Plot function for S3 class "atlasqtl"}
\usage{
\method{plot}{atlasqtl}(
  x,
  thres_ppi = 0.5,
  pch = 20,
  ylim_max = NULL,
  main = "Hotspot sizes",
  xlab = "Predictors",
  ylab = "sum_k PPI_st > thres",
  add = FALSE,
  ...
)
}
\arguments{
\item{x}{an object with S3 class \code{"atlasqtl"}}

\item{thres_ppi}{threshold to be applied on the posterior probabilities of 
association to define hotspots (default is 0.5, for the 
Median Probability Model).}

\item{pch}{type of points.}

\item{ylim_max}{upper limit for y-axis (must be set to NULL if \code{add} is TRUE).}

\item{main}{plot title (must be set to NULL if \code{add} is TRUE).}

\item{xlab}{x-axis title (must be set to NULL if \code{add} is TRUE).}

\item{ylab}{y-axis title (must be set to NULL if \code{add} is TRUE).}

\item{add}{whether the plot should be overlaid on an existing plot.}

\item{...}{additional arguments.}
}
\description{
This function provides functionality for plotting a so-called Manhattan plot 
with the position and size of "hotspot" predictors (i.e., predictors 
associated with multiple responses).
}
\seealso{
\code{\link{atlasqtl}}
}
