% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_output.R
\name{print.atlasqtl}
\alias{print.atlasqtl}
\title{Print function for S3 class "atlasqtl"}
\usage{
\method{print}{atlasqtl}(x, ...)
}
\arguments{
\item{x}{an object with S3 class \code{"atlasqtl"}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This function provides functionality for printing basic information about the 
atlasqtl run.
}
\seealso{
\code{\link{atlasqtl}}
}
