% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_output.R
\name{summary.atlasqtl}
\alias{summary.atlasqtl}
\title{Summary function for S3 class "atlasqtl"}
\usage{
\method{summary}{atlasqtl}(object, ...)
}
\arguments{
\item{object}{an object with S3 class \code{"atlasqtl"}}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
This function provides functionality for printing a basic summary of the 
atlasqtl posterior output.
}
\seealso{
\code{\link{atlasqtl}}
}
