% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlasqtl_package.R
\docType{package}
\name{atlasqtl-package}
\alias{atlasqtl-package}
\title{atlasqtl: a package for fAsT gLobal-locAl hotSpot QTL detection}
\description{
Global-local hierachical approach for joint detection of hotspot predictors 
in multiple-response regression. Inference uses closed-form variational 
updates coupled with a simulated annealing algorithm to enhance exploration 
of highly multimodal spaces. This software allows joint inference at 
large-scale, e.g., for molecular quantitative trait locus (QTL) studies where 
the hotspots are genetic variants regulating several molecular traits 
simultaneously. See Helene Ruffieux, Anthony C. Davison, Jorg Hager, 
Jamie Inshaw, Benjamin P. Fairfax, Sylvia Richardson, Leonardo Bottolo, A 
global-local approach for detecting hotspots in multiple-response regression, 
arXiv:1811.03334, 2018.
}
\section{atlasqtl functions}{
 atlasqtl, map_hyperprior_elicitation, set_hyper, set_init.
}

