% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_output.R
\name{plot.atlasqtl}
\alias{plot.atlasqtl}
\title{Plot function for S3 class "atlasqtl"}
\usage{
\method{plot}{atlasqtl}(
  x,
  thres = 0.5,
  fdr_adjust = FALSE,
  pch = 20,
  ylim_max = NULL,
  main = "Hotspot manhattan plot",
  xlab = "Predictors",
  ylab = "Nb of associated responses",
  add = FALSE,
  ...
)
}
\arguments{
\item{x}{an object with S3 class \code{"atlasqtl"}.}

\item{thres}{threshold to be applied on the posterior probabilities of 
association to define hotspots (default is 0.5, for the 
Median Probability Model).}

\item{fdr_adjust}{if TRUE the threshold is applied on FDR estimates computed
from the "atlasqtl" posterior probabilities of association,
otherwise it is directly applied on the posterior 
probabilities of association (default is FALSE).}

\item{pch}{type of points.}

\item{ylim_max}{upper limit for y-axis (must be set to NULL if \code{add} is TRUE).}

\item{main}{plot title (must be set to NULL if \code{add} is TRUE).}

\item{xlab}{x-axis title (must be set to NULL if \code{add} is TRUE).}

\item{ylab}{y-axis title (must be set to NULL if \code{add} is TRUE).}

\item{add}{whether the plot should be overlaid on an existing plot.}

\item{...}{additional arguments.}
}
\description{
This function provides functionality for plotting a so-called Manhattan plot 
with the position and size of "hotspot" predictors (i.e., predictors 
associated with multiple responses).
}
\seealso{
\code{\link{atlasqtl}}
}
