% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_hyper_init.R
\name{set_init}
\alias{set_init}
\title{Gather initial variational parameters provided by the user.}
\usage{
set_init(q, p, gam_vb, mu_beta_vb, sig2_beta_vb, tau_vb)
}
\arguments{
\item{q}{Number of responses.}

\item{p}{Number of candidate predictors.}

\item{gam_vb}{Matrix of size p x q with initial values for the variational
parameter yielding posterior probabilities of inclusion.}

\item{mu_beta_vb}{Matrix of size p x q with initial values for the
variational parameter yielding regression coefficient estimates for
predictor-response pairs included in the model.}

\item{sig2_beta_vb}{Vector of length q. For these values are the same
for all the predictors (as a result of the predictor variables being
standardized before running the variational algorithm).}

\item{tau_vb}{Vector of length q with initial values for the variational 
parameter yielding estimates for the continuous response residual 
precisions.}
}
\value{
An object of class "\code{init}" preparing user initial values for
  the variational parameters in a form that can be passed to the
  \code{\link{atlasqtl}} function.
}
\description{
This function must be used to provide initial values for the variational
parameters used in \code{\link{atlasqtl}}.
}
\details{
The \code{\link{atlasqtl}} function can also be used with default initial
parameter choices (without using \code{\link{set_init}}) by setting
its argument \code{list_init} to \code{NULL}.
}
\examples{
seed <- 123; set.seed(seed)

###################
## Simulate data ##
###################

## Examples using small problem sizes:
##
n <- 200; p <- 200; p_act <- 20; q <- 20; q_act <- 15

## Candidate predictors (subject to selection)
##
# Here we simulate common genetic variants (but any type of candidate
# predictors can be supplied).
# 0 = homozygous, major allele, 1 = heterozygous, 2 = homozygous, minor allele

X_act <- matrix(rbinom(n * p_act, size = 2, p = 0.25), nrow = n)
X_inact <- matrix(rbinom(n * (p - p_act), size = 2, p = 0.25), nrow = n)

shuff_x_ind <- sample(p)
X <- cbind(X_act, X_inact)[, shuff_x_ind]

bool_x_act <- shuff_x_ind <= p_act

pat_act <- beta <- matrix(0, nrow = p_act, ncol = q_act)
pat_act[sample(p_act*q_act, floor(p_act*q_act/5))] <- 1
beta[as.logical(pat_act)] <-  rnorm(sum(pat_act))

## Gaussian responses
##
Y_act <- matrix(rnorm(n * q_act, mean = X_act \%*\% beta, sd = 0.5), nrow = n)
Y_inact <- matrix(rnorm(n * (q - q_act), sd = 0.5), nrow = n)
shuff_y_ind <- sample(q)
Y <- cbind(Y_act, Y_inact)[, shuff_y_ind] 


########################
## Infer associations ##
########################

## Continuous responses
##

# gam_vb chosen so that the prior mean number of responses associated with
# each candidate predictor is 1/4.
gam_vb <- matrix(rbeta(p * q, shape1 = 1, shape2 = 4*q-1), nrow = p)
mu_beta_vb <- matrix(rnorm(p * q), nrow = p)
tau_vb <- 1 / apply(Y, 2, var)
sig2_beta_vb <- 1 / rgamma(q, shape = 2, rate = 1 / tau_vb)

list_init <- set_init(q, p, gam_vb, mu_beta_vb, sig2_beta_vb, tau_vb)

vb_g <- atlasqtl(Y = Y, X = X, p0 = c(5, 25), list_init = list_init)

}
\seealso{
\code{\link{set_hyper}}, \code{\link{atlasqtl}}
}
