% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_output.R
\name{summary.atlasqtl}
\alias{summary.atlasqtl}
\title{Summary function for S3 class "atlasqtl"}
\usage{
\method{summary}{atlasqtl}(object, thres, fdr_adjust = FALSE, ...)
}
\arguments{
\item{object}{an object with S3 class \code{"atlasqtl"}}

\item{thres}{threshold to be applied on the posterior probabilities of 
association (PPI) to define hotspots (default is 0.5, for the 
Median Probability Model).}

\item{fdr_adjust}{if TRUE the threshold is applied on FDR estimates computed
from the "atlasqtl" posterior probabilities of association,
otherwise it is directly applied on the posterior 
probabilities of association (default is FALSE).}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
This function provides functionality for printing a basic summary of the 
atlasqtl posterior output.
}
\seealso{
\code{\link{atlasqtl}}
}
